Add comprehensive tests for exam validation and submission processes
- Implement integration tests for reading, listening, and writing exam submissions, covering validation failures and successful submissions.
- Create unit tests for exam timer calculations and submission validation, ensuring proper handling of edge cases and error scenarios.
- Introduce rate limiting tests for authentication and password reset functionalities.
- Validate independent offsets for exam retrieval and ensure correct handling of invalid test references.
- Enhance the writing submission process to handle AI evaluation failures and admin review scenarios.

+++ backend/app/modules/exams/schemas/models.py
@@ -24,7 +24,7 @@ class ExamSubmitOut(BaseModel):
     answers: list[dict[str, Any]]
     score: float | None
     correct_answers: int | None
-    time_spent: float | None
+    time_spent: int | None
 
 
 class ExamPublic(BaseModel):
